# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  aslon1213_gift_pkg_repository.Empty:
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.Empty'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_User:
    properties:
      data:
        $ref: '#/definitions/aslon1213_gift_pkg_repository.User'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
  aslon1213_gift_pkg_repository.Response-pkg_handlers_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/pkg_handlers.LoginResponse'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
  aslon1213_gift_pkg_repository.Response-pkg_handlers_RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/pkg_handlers.RefreshResponse'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
  aslon1213_gift_pkg_repository.Response-pkg_handlers_RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/pkg_handlers.RegisterResponse'
      message:
        example: ok
        type: string
      status:
        example: success
        type: string
    type: object
...
  aslon1213_gift_pkg_repository.User:
    properties:
      balance:
        type: number
      created_at:
        type: string
      currency:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
...
  pkg_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  pkg_handlers.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  pkg_handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  pkg_handlers.RefreshResponse:
    properties:
      refresh_token:
        type: string
      token:
        type: string
    type: object
  pkg_handlers.RegisterRequest:
    properties:
      currency:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  pkg_handlers.RegisterResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: hamidovaslon1@gmail.com
    name: API Support
    url: https://github.com/aslon1213/gift
  description: This is a Gift API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/aslon1213/gift
  title: Gift API
  version: 0.1.0
paths:
...
  /api/v1/alerts/{id}:
    delete:
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Delete alert
      tags:
      - alerts
    get:
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Get alert by ID
      tags:
      - alerts
    put:
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Update alert
      tags:
      - alerts
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Retrieves the authenticated user's information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Get user info
      tags:
      - auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in with email and password
      parameters:
      - description: Login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-pkg_handlers_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      description: Log out and revoke all refresh tokens for user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      security:
      - ApiKeyAuth: []
      summary: Log out
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access and refresh tokens using a refresh token
      parameters:
      - description: Refresh request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-pkg_handlers_RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Refresh JWT tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Registration request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/pkg_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-pkg_handlers_RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/aslon1213_gift_pkg_repository.Response-aslon1213_gift_pkg_repository_Empty'
      summary: Register user
      tags:
      - auth
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"